genai_chatbot_hackathon/requirements.txt
Flask
openai
pinecone-client
streamlit
azure-search-documents
python-dotenv